+++ src/main/java/persistence/Application.java
@@ -6,10 +6,10 @@ import jdbc.JdbcTemplate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import persistence.entity.MyEntityManager;
-import persistence.sql.Person;
+import domain.Person;
 import persistence.sql.ddl.CreateQueryBuilder;
 import persistence.sql.ddl.DropQueryBuilder;
-import persistence.sql.domain.dialect.H2Dialect;
+import database.dialect.H2Dialect;
 import repository.CustomJpaRepository;
 
 public class Application {
