+++ src/main/java/persistence/sql/dml/InsertQueryBuilder.java
@@ -1,8 +1,8 @@
 package persistence.sql.dml;
 
-import persistence.sql.domain.Column;
-import persistence.sql.domain.DataType;
-import persistence.sql.domain.Table;
+import persistence.sql.meta.Column;
+import persistence.sql.meta.DataType;
+import persistence.sql.meta.Table;
 import utils.ValueExtractor;
 
 import java.util.LinkedHashMap;
