+++ src/main/java/persistence/sql/dml/SelectAllQueryBuilder.java
@@ -1,7 +1,7 @@
 package persistence.sql.dml;
 
-import persistence.sql.domain.Column;
-import persistence.sql.domain.Table;
+import persistence.sql.meta.Column;
+import persistence.sql.meta.Table;
 
 import java.util.List;
 import java.util.stream.Collectors;
