+++ src/main/java/persistence/sql/meta/DataType.java
@@ -1,4 +1,4 @@
-package persistence.sql.domain;
+package persistence.sql.meta;
 
 import persistence.exception.NotAllowedDataTypeException;
