+++ src/test/java/persistence/entity/MyEntityLoaderTest.java
@@ -4,10 +4,10 @@ import jdbc.JdbcTemplate;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import persistence.sql.Person;
+import domain.Person;
 import persistence.sql.ddl.CreateQueryBuilder;
 import persistence.sql.ddl.DropQueryBuilder;
-import persistence.sql.domain.dialect.H2Dialect;
+import database.dialect.H2Dialect;
 import persistence.support.DatabaseSetup;
 
 import static org.assertj.core.api.Assertions.assertThat;
