+++ src/test/java/persistence/entity/MyEntityPersisterTest.java
@@ -4,11 +4,11 @@ import jdbc.JdbcTemplate;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import persistence.sql.Person;
+import domain.Person;
 import persistence.sql.ddl.CreateQueryBuilder;
 import persistence.sql.ddl.DropQueryBuilder;
 import persistence.sql.dml.SelectAllQueryBuilder;
-import persistence.sql.domain.dialect.H2Dialect;
+import database.dialect.H2Dialect;
 import persistence.support.DatabaseSetup;
 
 import java.util.List;
