+++ src/test/java/persistence/sql/ddl/CreateQueryBuilderTest.java
@@ -4,8 +4,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import persistence.exception.NotEntityException;
-import persistence.sql.Person;
-import persistence.sql.domain.dialect.H2Dialect;
+import domain.Person;
+import database.dialect.H2Dialect;
 import persistence.study.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
