+++ src/test/java/persistence/sql/dml/SelectQueryBuilderTest.java
@@ -2,7 +2,7 @@ package persistence.sql.dml;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import persistence.sql.Person;
+import domain.Person;
 
 import static org.assertj.core.api.Assertions.assertThat;
