+++ src/test/java/persistence/sql/dml/UpdateQueryBuilderTest.java
@@ -1,9 +1,8 @@
 package persistence.sql.dml;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import persistence.sql.Person;
+import domain.Person;
 
 import static org.assertj.core.api.Assertions.*;
